requirements.txt
discord.py
python-dotenv
mathparse
PyNaCl